Error Handler Relates #24
